pyproject.toml
[project]
name = "balance-news"
version = "0.1.0"
description = "null"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.2.1",
    "reportlab>=4.4.1",
    "svglib>=1.5.1",
]
